+++ framework/src/main/kotlin/me/wieku/framework/animation/GliderEvent.kt
@@ -3,8 +3,8 @@ package me.wieku.framework.animation
 import me.wieku.framework.math.Easing
 
 data class GliderEvent(
-    val startTime: Float,
-    val endTime: Float,
+    val startTime: Double,
+    val endTime: Double,
     val startValue: Float,
     val endValue: Float,
     val easing: Easing
