+++ framework/src/main/kotlin/me/wieku/framework/time/FrameInfo.kt
@@ -1,3 +1,3 @@
 package me.wieku.framework.time
 
-data class FrameInfo(var currentTime: Float, var lastTime: Float, var frameTime: Float)
+data class FrameInfo(var currentTime: Double, var lastTime: Double, var frameTime: Double)
